[docker-snmp] Provides a config file for the polling interval in sonic_ax_impl
By default, SONiC does not restrict the CPU usage of the system containers.
This caused the SNMP container to consume 100% CPU.

A feasible approach is to define a configure file to adjust the CPU usage of the
SNMP container when it is required under specific scenarios.
The interval of the polling tasks in the SNMP process also affects the fundamental
CPU usage consumed by the SNMP process. So, the configure file also needs to provide
the option to change the update intervals in the SNMP process.